iotivity-node: pass arm64 to npm for aarch64
[GDP-571] Iotivity build failure on Dragonboard

In iotivity-node<version>.bb, do_compile() includes a case statement to
translate some known hardware architectures to the strings used when
passing --arch flag to the npm command.  The compilation target
architecture of Dragonboard (and several other 64 bit ARM boards) is
aarch64. aarch64 was not included in the known list, resulting in
the printout "unknown architecture" and stopped operation.

Add "aarch64" and translate it to what npm expects, which is "arm64".

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>